requirements.txt
Authlib==1.4.1
Django==4.2.19
PyJWT==2.10.1
asgiref==3.8.1
base-django-rest-framework==1.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
django-allauth==65.4.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
hiredis==2.4.0
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
pip==23.2.1
psycopg==3.2.4
pycparser==2.22
python3-openid==3.2.0
redis==4.6.0
requests==2.32.3
requests-oauthlib==2.0.0
sentry-sdk==1.45.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
